Move parse_datetime to utils
