+++ tests/test_io_utils.py
@@ -15,8 +15,8 @@ from io_utils import (
     write_summary,
     copy_config,
     apply_burst_filter,
-    parse_datetime,
 )
+from utils import parse_datetime
 
 
 def test_load_config(tmp_path):
